Update for the the TI AM437x EVAL board.
We support 2 different build configurations now:
 - am437x_evm_defconfig - builds a UBIFS root filsystem for NAND flash
 - am437x_evm_recovery_defconfig - builds a small kernel+RAMFS suitable
for update/recovery use.

Both builds now support using swupdate to update the running firmware in
NAND Flash. This is using a double copy strategy with fall back.

To support the double copy strategy, the NAND partitioning was changed
to match that Fremont desired. This requires patches to both the u-boot
and kernel source files from the default TI source base. The patches are
integrated into the build of both.

Building of the new image file, Pocket-30_1.0.1.swu, is done as the last
step of the buildroot process. It takes the output images from the build
and produces the Pocket-30_1.0.1.swu image which is suitable for
updating the NAND flash on the TI AM437x eval board right now.

The special "recovery" image was developed as an alternative to the
dual copy strategy and uses a method similar to Android recovery mode
where the device boots into a RAM-based image that allows updating any
partition on the flash device as necessary. In order to accomplish this
over the USB port, a special USB MTP (Media Transfer Protocol)
Responder daemon, umtpd, was added that presents a USB mass storage
device to the attached USB Host device where the .swu file can be
transferred and the swupdate process is automatically launched.