Update u-boot with new environment variables and enable the BOOT_COUNT option which will run the "altboot" command if "bootlimit" value, currently set to 3 is reached. This prevents a device from constantly rebooting of the image fails to boot properly. On the TI amxxx devices, the boot count value that u-boot sets/reads is stored to the CPU's RTC Scratch2 register which is otherwise unused and can survive reboots.
We also modify how the partition to be made active (primary or
secondary images). We now created a u-boot command, switchactive, which
will be called if booting fails to switch to alternate boot image based
on which one is currently active.

The currently active image is selected via new environment variable
bootselect which is either not present (boots primary) or set to "2"
which boots the alternate/secondary image.

To set the correct image a new u-boot command, nandselectroot, will
configure the nandroot variable correctly.