With new u-boot support added for the BOOT_COUNT option when Linux boots we need to reset the boot counter for subsequent boot operations so we don't trigger the bootlimit setting.
On the TI AMxxx u-boot builds, the boot count value is saved to the RTC
Scratch2 register on the CPU itself which will survice reboot.

NOTE: that in order to write to the RTC registers it's necessary to
first write 2 special values in order to unlock them otherwise writing
is locked out.